Remove thesis title from page titles (#121)
It got a bit confusing. Use the degree name instead. Set thesis title as
metadata and insert in info sidebar. This allows using the title and
thesis title in the about page as well. Strip formatting from titles
when setting the "title" in the header.
Add sucupira entry for PhD thesis.