Simplify the menu and redesign the front page (#227)
Add the TL;DR section to the front page with link to the CompGeoLab.
Remove my name from the menu and replace with "Home".
Flatten the menu into a single column and add more vertical padding.
Replace slogan in the front page with my name and keywords.
Make the banner images slightly darker for more contrast.