Fix antialias issue in banner noise and make them smaller
The issue with using a single noise pixel and tiling it was the
antialiasing that the browser was or wasn't doing at different zoom
levels. Made them into a tiled noise image and that problem went away.
Reduce the size of all banners to save space, move them to
images/banners for easier time editing them in the future.